requirements.txt
django==2.2.5
mysqlclient==2.0.1
pillow==7.2.0
requests==2.24.0
django-simple-captcha
redis==3.5.3
aliyun-python-sdk-core==2.13.25
aliyun-python-sdk-ecs==4.19.10
django-pure-pagination==0.3.0
